refactor scheme compiler
